test: package 정리
