+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.Team;
 import chess.domain.piece.Bishop;
 import chess.domain.position.Position;
 import org.assertj.core.api.Assertions;
